resolve the issue of extending boundary
i set a delayed mean and compare the reading pixel to the delayed mean
of the likely foregrounds. it seems that the delay successfully fixed
the instability issue of the 30% cutoff previously, that some images may
have smoothed boundaries. now, the delay is set to 10 px, and may be
enough for us. i do not make this a parameter, since this may not change
that frequently.